Move swan objects that will be deprecated from swan config into legacy module
